+++ includes/DependencyManagement/Providers/AdminDashboardServiceProvider.php
@@ -4,6 +4,7 @@ namespace WeDevs\Dokan\DependencyManagement\Providers;
 
 use WeDevs\Dokan\Admin\Dashboard\Dashboard;
 use WeDevs\Dokan\Admin\Dashboard\Pages\Modules;
+use WeDevs\Dokan\Admin\Dashboard\Pages\ProFeatures;
 use WeDevs\Dokan\Admin\Dashboard\Pages\Settings;
 use WeDevs\Dokan\Admin\Dashboard\Pages\SetupGuide;
 use WeDevs\Dokan\Admin\Dashboard\Pages\Status;
@@ -11,22 +12,37 @@ use WeDevs\Dokan\Admin\OnboardingSetup\AdminSetupGuide;
 use WeDevs\Dokan\DependencyManagement\BaseServiceProvider;
 use WeDevs\Dokan\Admin\Settings\Settings as AdminSettingsProvider;
 
+/**
+ * Admin Dashboard API Service Provider
+ *
+ * @since 4.1.0
+ */
 class AdminDashboardServiceProvider extends BaseServiceProvider {
     /**
      * Tag for services added to the container.
+     *
+     * @var array
      */
     protected $tags = [ 'admin-dashboard-service' ];
 
-	protected $services = [
+    /**
+     * Services to register
+     *
+     * @var array
+     */
+    protected $services = [
         Dashboard::class,
         Modules::class,
         Status::class,
+        ProFeatures::class,
 	];
 
-	/**
-     * Register the classes.
+    /**
+     * Register the services.
+     *
+     * @return void
      */
-	public function register(): void {
+    public function register(): void {
         foreach ( $this->services as $service ) {
             $definition = $this->share_with_implements_tags( $service );
             $this->add_tags( $definition, $this->tags );