+++ includes/DependencyManagement/Providers/ServiceProvider.php
@@ -62,9 +62,9 @@ class ServiceProvider extends BootableServiceProvider {
 		$this->getContainer()->addServiceProvider( new FrontendServiceProvider() );
 		$this->getContainer()->addServiceProvider( new AjaxServiceProvider() );
 		$this->getContainer()->addServiceProvider( new AnalyticsServiceProvider() );
-    $this->getContainer()->addServiceProvider( new UtilsServiceProvider() );
+        $this->getContainer()->addServiceProvider( new UtilsServiceProvider() );
 		$this->getContainer()->addServiceProvider( new CommissionServiceProvider() );
-    $this->getContainer()->addServiceProvider( new IntelligenceServiceProvider() );
+        $this->getContainer()->addServiceProvider( new IntelligenceServiceProvider() );
 		$this->getContainer()->addServiceProvider( new AdminDashboardServiceProvider() );
 		$this->getContainer()->addServiceProvider( new AdminSetupGuideServiceProvider() );
 		$this->getContainer()->addServiceProvider( new ModelServiceProvider() );